--- 307-big-ballers/src/app/index.js
@@ -1,2 +1,0 @@
-
-import dotenv from "dotenv";
